+++ src/Routers/Routers.jsx
@@ -1,11 +1,11 @@
 import { Routes, Route, BrowserRouter as Router } from 'react-router-dom';
 import AllListPage from '../pages/AllListPage/AllListPage';
 import ListPage from '../pages/ListPage/ListPage';
-import RollingPaper from '../pages/RollingPaper';
+import RollingPaper from '../pages/RollingPaperPage/RollingPaperPage';
 import Header from '../components/common/Header/Header';
 import CreateTo from '../components/create/CreateTo';
 import MainPage from '../pages/MainPage/MainPage';
-import MessageAddPage from '../pages/MessageAddPage';
+import MessageAddPage from '../pages/MessageAddPage/MessageAddPage';
 
 function Routers() {
   return (
