+++ src/components/RollingPaper/Avatar/Avatar.styled.js
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const Image = styled.img`
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+`;
