+++ src/components/RollingPaper/DropDown/DropDown.styled.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import colors from '../../styles/colors';
+import colors from '../../../styles/colors';
 
 const BaseDropDown = styled.div`
   border-radius: 8px;
