+++ src/components/RollingPaper/MessageCardList/MessageCardList.jsx
@@ -1,27 +1,11 @@
-import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
-import MessageCard from './MessageCard/MessageCard';
-import PlusCard from './MessageCard/PlusCard';
-
-const List = styled.ul`
-  width: 100%;
-  max-width: 1200px;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 20px;
-
-  @media (max-width: 820px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (max-width: 475px) {
-    grid-template-columns: repeat(1, 1fr);
-  }
-`;
+import MessageCard from '../MessageCard/MessageCard/MessageCard';
+import PlusCard from '../MessageCard/PlusCard/PlusCard';
+import * as S from './MessageCardList.styled';
 
 function CardList({ messages, lastRef, onClick, isEditPage }) {
   return (
-    <List>
+    <S.List>
       {!isEditPage && (
         <Link to="message">
           <PlusCard />
@@ -33,7 +17,7 @@ function CardList({ messages, lastRef, onClick, isEditPage }) {
           <div ref={idx === messages.length - 3 ? lastRef : null} />
         </li>
       ))}
-    </List>
+    </S.List>
   );
 }
