+++ src/components/RollingPaper/MessageCardList/MessageCardList.styled.js
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const List = styled.ul`
+  width: 100%;
+  max-width: 1200px;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 20px;
+
+  @media (max-width: 820px) {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  @media (max-width: 475px) {
+    grid-template-columns: repeat(1, 1fr);
+  }
+`;
