+++ src/components/RollingPaper/TextViewer/TextViewer.jsx
@@ -1,19 +1,12 @@
 import { Viewer } from '@toast-ui/react-editor';
 import '@toast-ui/editor/dist/toastui-editor-viewer.css';
-import styled from '@emotion/styled';
-
-const Message = styled.div`
-  .toastui-editor-contents {
-    font-size: 1.125rem;
-    font-family: ${({ font }) => font};
-  }
-`;
+import * as S from './TextViewer.styled';
 
 function TextViewer({ message, font }) {
   return (
-    <Message font={font}>
+    <S.Message font={font}>
       <Viewer initialValue={message || ''} />
-    </Message>
+    </S.Message>
   );
 }
