+++ src/components/RollingPaper/TextViewer/TextViewer.styled.js
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const Message = styled.div`
+  .toastui-editor-contents {
+    font-size: 1.125rem;
+    font-family: ${({ font }) => font};
+  }
+`;
