+++ src/components/RollingPaper/ToastBar/ToastBar.jsx
@@ -1,0 +1,7 @@
+import * as S from './ToastBar.styled';
+
+function ToastBar({ children }) {
+  return <S.Toast>{children}</S.Toast>;
+}
+
+export default ToastBar;
